// src/app/lib/riot/types.ts
export type TeamId = 100 | 200;

export interface Summoner {
  platformId: string;
  accountId: string;
  profileIconId?: number;
  revisionDate: number;
  name: string;
  id: string;
  puuid: string;
  summonerLevel: number;
}

export type Position = {
  x: number;
  y: number;
};
// ...
export type EliteMonsterKill = {
  assistingParticipantIds?: number[];
  killerId: number;
  killerTeamId: TeamId;
  monsterSubType?:
    | 'FIRE_DRAGON'
    | 'EARTH_DRAGON'
    | 'AIR_DRAGON'
    | 'WATER_DRAGON'
    | 'ELDER_DRAGON';
  monsterType: 'RIFTHERALD' | 'DRAGON' | 'BARON_NASHOR';
  position: Position;
  timestamp: number;
  type: 'ELITE_MONSTER_KILL';
};